show tour deatail on different page
