+++ src/components/TourDetail.js
@@ -1,15 +1,34 @@
-import React from 'react';
+import React, { useEffect, useState } from 'react';
+import { useDispatch, useSelector } from 'react-redux';
 import { useParams } from 'react-router-dom';
 import Dropdown from 'react-bootstrap/Dropdown';
+import getTourDetail from '../redux/services/tourdetail.service';
 import imgExample from '../img/tour1.jpg';
 
 function TourDetail() {
+  const dispatch = useDispatch();
+  const { tourdetail } = useSelector(state => state.tourdetailStore);
+  const { id } = useParams();
+  const [tourDetailReceived, settourDetailReceived] = useState([]);
+
+  useEffect(() => {
+    dispatch(getTourDetail(id));
+  }, [id]);
+
+  useEffect(() => {
+    if (tourdetail) {
+      settourDetailReceived(tourdetail);
+    }
+  }, [tourdetail]);
+
   return (
     <div>
-      <img src={imgExample} alt="images" />
-      <h2>Titulo del tours  Titulo del tours Titulo del tours </h2>
-      <p>lorem askdj askdjaskdjalskdj asd jasdlkj asldk jasdk jaslkd jasldk jaslkd j lorem askdj askdjaskdjalskdj asd jasdlkj asldk jasdk jaslkd jasldk jaslkd j</p>
-      <h1>$17.00</h1>
+      <img src={tourDetailReceived.pictures} alt="images" />
+      <h2>
+        { tourDetailReceived.name }
+      </h2>
+      <p>{tourDetailReceived.shortDescription}</p>
+      <h1> </h1>
       <p>Quantity</p>
       <Dropdown>
         <Dropdown.Toggle variant="success" id="dropdown-basic">
